Add support for static libs
Static libraries can be linked with the regular "lib" attribute, but this will make symbols of a static library public. When using the static_lib attribute, the static lib will first be unpacked into separate object files, and then those object files will be linked with individually. That allows for full control of symbol visibility.